feat: implement JWT auth middleware and protect note routes

+++ backend/routes/noteRoutes.js
@@ -1,12 +1,12 @@
 import express, { json } from 'express'
 import Note from '../models/note.js'
-
+import auth from '../middleware/auth.js'
 const router=express.Router()
 
-router.post('/',async(req,res)=>{
+router.post('/',auth,async(req,res)=>{
     try{
         const {title,description}=req.body;
-        const newNote=new Note({title,description})
+        const newNote=new Note({title,description,user:req.user.id})
         const savedNote = await newNote.save();
         res.status(201).json(savedNote)
     }catch(error){
@@ -25,10 +25,14 @@ router.get('/',async (req,res)=>{
 
 router.delete('/:id',async (req,res)=>{
     try{
-        const deletedNote=await Note.findByIdAndDelete(req.params.id)
+        const note=await Note.findById(req.params.id)
         if(!deletedNote){
             return res.status(404).json({message:"Note not found"})
         }
+        if(note.user.toString!=req.user.id){
+            return res.status(400).json({message :"Not authorised to delete this note"})
+        }
+        await Note.findByIdAndDelete(req.params.id)
         res.status(200).json({message:"Note deleted Successfully",deletedNote})
     }catch(error){
         return res.status(500).json({Error: error.message})
@@ -38,10 +42,18 @@ router.delete('/:id',async (req,res)=>{
 router.put('/:id',async (req,res)=>{
     try{
         const {title,description}=req.body;
-        const updatedNote=await Note.findByIdAndUpdate(req.params.id,{title,description},{new: true})
-        if(!updatedNote){
+        const note=await Note.findById(req.params.id,{title,description},{new: true})
+        if(!note){
             return res.status(404).json({message:"Note not found"})
         }
+        if(note != req.user.id){
+            return res.status(400).json({message : "Not authorised to update this note"})
+        }
+        const updatedNote = await Note.findByIdAndUpdate(
+            req.params.id,
+            {title,description},
+            {new:true}
+        )
         res.status(200).json({message:"Note Updated Successfully",updatedNote})
     }catch(error){
         res.status(500).json({Error: error.message})
